std::call_once: fix a link
It might seem like it's been broken since forever, but it's actually a
consequence of my idiotic idea to completely rewrite this repository's
history.
One of the things I wanted to fix is I would like the code to reside in
master.
I did import it, but messed up the history completely as a result.