fix navbar links with multiple pages
I determine whether a page should be present in the navigation bar by
checking if 'navbar' is present in this page's `groups`.
My index.html is an example of such page.
But `jekyll-paginate` creates a bunch of similar pages, each of them
having 'navbar' in their `groups`.
Thus, every /pageN appears in the navbar.
To get rid of all of them except for the first one, I set `root_page` to
`true` in index.html, and only include the first page with `root_page ==
true` in the navbar.
Man, it's a hack if I've ever seen one.